Google Sign In with Firebase fix (plus default photo fix)
The Google login used to work, but authentications were not updated in the Firebase project. I changed the way this authentication is done, using the way specified in Firebase -> Authenticate using Google mixed with my previous solution (GoogleSignInClient, GoogleSignInOptions...) and now the users appear in the Authentication section in my Firebase project.

This also fixed the default profile picture not working. If an account has the default Google letter picture my previous solution showed my own default picture instead because the photoUrl returned null, but it now shows the correct default image (for example, it used to show "usuario.png" with an account starting with D and no photo, but now it shows the "D" image with a colored background)